internal/randvar: take uint64 in IncMax
Take a uint64, not an int, in (randvar.Dynamic).IncMax to match the interface's
other methods.

+++ internal/randvar/randvar.go
@@ -18,7 +18,7 @@ type Dynamic interface {
 	Static
 
 	// Increment the max value the variable will return.
-	IncMax(delta int)
+	IncMax(delta uint64)
 
 	// Read the current max value the variable will return.
 	Max() uint64
